fix: resolve clang-tidy implicit widening warnings
Fixes 3 clang-tidy warnings about implicit widening conversions
in discovery.c:

1. ADDR_FRESH_THRESHOLD_SEC: Changed from (3 * 60 * 60) to (3ULL * 60 * 60)
2. ADDR_MAX_AGE_SEC: Changed from (7 * 24 * 60 * 60) to (7ULL * 24 * 60 * 60)
3. Score calculation: Cast entry->attempts to uint64_t before multiplication

These constants are compared with uint64_t values, so they need to be
explicitly typed as unsigned long long to avoid implicit widening.

All tests still pass (20/20).

Addresses bugprone-implicit-widening-of-multiplication-result warnings.

+++ src/protocol/discovery.c
@@ -3,6 +3,7 @@
  */
 
 #include "discovery.h"
+#include "echo_types.h"
 #include "platform.h"
 #include "protocol.h"
 #include <stdint.h>
@@ -100,8 +101,8 @@ static echo_bool_t is_unspecified(const uint8_t *ipv6) {
   }
 
   /* Check for IPv4-mapped 0.0.0.0 (::ffff:0.0.0.0) */
-  if (is_ipv4_mapped(ipv6) &&
-      ipv6[12] == 0 && ipv6[13] == 0 && ipv6[14] == 0 && ipv6[15] == 0) {
+  if (is_ipv4_mapped(ipv6) && ipv6[12] == 0 && ipv6[13] == 0 && ipv6[14] == 0 &&
+      ipv6[15] == 0) {
     return ECHO_TRUE;
   }
 
@@ -419,7 +420,7 @@ echo_result_t discovery_select_outbound_address(peer_addr_manager_t *manager,
 
     /* Penalty for failed attempts */
     if (entry->attempts > 0) {
-      score -= entry->attempts * 10000;
+      score -= (uint64_t)entry->attempts * 10000;
     }
 
     /* Consider this address if it's the best so far */
